+++ src/rule-resolver/filter-rule-definitions.ts
@@ -1,0 +1,19 @@
+import { RuleDefinitions } from './../types';
+
+export default (rules: RuleDefinitions, filter: string[] | undefined): RuleDefinitions => {
+    if (!filter) {
+        return rules;
+    }
+
+    const result = Object.entries(rules)
+        .filter(([ruleName]) => filter.includes(ruleName))
+        .reduce(
+            (acc, [ruleName, ruleDefinition]) => {
+                acc[ruleName] = ruleDefinition;
+                return acc;
+            },
+            {} as RuleDefinitions,
+        );
+
+    return result;
+};