+++ src/rule-resolver/index.ts
@@ -1,0 +1,16 @@
+import { ConfigRules, RuleApplications } from './../types';
+import getRuleApplications from './get-rule-applications';
+import filterRuleDefinitions from './filter-rule-definitions';
+import getRuleDefinitions from './get-rule-definitions';
+
+export default (
+    rules: ConfigRules,
+    rulesDir: string | undefined,
+    rulesToVerify: string[] | undefined,
+): RuleApplications => {
+    const ruleDefinitions = getRuleDefinitions(rules, rulesDir);
+    const filteredRuleDefinitions = filterRuleDefinitions(ruleDefinitions, rulesToVerify);
+    const ruleApplications = getRuleApplications(rules, filteredRuleDefinitions, rulesToVerify);
+
+    return ruleApplications;
+};