+++ src/stricter.ts
@@ -1,13 +1,9 @@
 import { getConfig } from './config';
-import {
-    getRuleDefinitions,
-    getRuleApplications,
-    filterFilesToProcess,
-    filterRuleDefinitions,
-} from './rule';
-import { applyProjectRules, readFilesData } from './processor';
+import getRuleApplications from './rule-resolver';
+import getFileList from './file-resolver';
+import processFiles from './file-processor';
+import { applyProjectRules } from './processor';
 import { getErrorCount } from './reporter';
-import { listFiles } from './utils';
 import { StricterArguments } from './types';
 
 export default ({
@@ -25,23 +21,14 @@ export default ({
     debug('Read config');
     const config = getConfig(configPath);
 
+    debug('Get rules');
+    const ruleApplications = getRuleApplications(config.rules, config.rulesDir, rulesToVerify);
+
     debug('Get file list');
-    const fileList = listFiles(config.root, config.exclude);
-
-    debug('Get rule definitions');
-    const ruleDefinitions = getRuleDefinitions(config);
-
-    debug('Filter rule definitions');
-    const filteredRuleDefinitions = filterRuleDefinitions(ruleDefinitions, rulesToVerify);
-
-    debug('Get rule applications');
-    const ruleApplications = getRuleApplications(config, filteredRuleDefinitions, rulesToVerify);
-
-    debug('Get files to process');
-    const filesToProcess = filterFilesToProcess(config.root, fileList, ruleApplications);
+    const filesToProcess = getFileList(config.root, config.exclude, ruleApplications);
 
     debug('Read files data');
-    const filesData = readFilesData(filesToProcess);
+    const filesData = processFiles(filesToProcess);
 
     debug('Apply rules');
     const projectResult = applyProjectRules(config.root, filesData, ruleApplications);